+++ common/service/k8s/higress/higresscontroller_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/w7panel/w7panel/common/service/k8s"
 )
 
-func TestInitW7ProxyPlugin(t *testing.T) {
+func disabledInitW7ProxyPlugin(t *testing.T) {
 	os.Setenv("KO_DATA_PATH", "/home/workspace/k8s-offline/kodata")
 	client, err := k8s.NewK8sClient().Sdk.ToSigClient()
 	if err != nil {